# Performance: don't use VM when not needed