--- a/index.js
+++ b/index.js
@@ -8,7 +8,7 @@ function escape (template) {
 
 function compile (template, defaultContext, ops) {
   if (typeof template !== 'string') {
-    throw new Error('Template must be a string')
+    throw new TypeError('Template must be a string')
   }
   const options = Object.assign({timeout: 500}, ops)
   const script = new vm.Script(escape(template))