--- a/package.json
+++ b/package.json
@@ -24,6 +24,9 @@
   "engines": {
     "node": ">=4.0.0"
   },
+  "files": [
+    "index.js"
+  ],
   "devDependencies": {
     "ava": "^0.22.0",
     "eslint": "^4.4.1",