@dotcypress 2019-08-16
@EdJoPaTo I think to fix performance issue we need to patch `telegraf-i18n` not `compile-template`, because template library don't make any sense with `templateless` input 😉