@EdJoPaTo 2019-08-17
In my opinion the template without a function in it is the simplest form of a template.
implementing this on the user side off this lib (telegraf-i18n) does not make sense to me as they just have some kind of template. how to handle the template effectively is not the job of the user, its the job of the template engine. telegraf-i18n has the job of getting the right template on the given language and resourceKey so it shouldn't be the task of it to optimise performance for something it isn't made for.

Everyone that may want to use this library might want the same: fast templates with every kind of template they stick into it. But they have to do the same kind of optimisation which telegraf-i18n does again to use this library more efficiently.

When asking the question "what's the purpose of telegraf-i18n? What does it abstract for me?" it does not make sense to me to change that there.